+++ src/components/projects/projectId/playground/PlaygroundSearch.tsx
@@ -2,20 +2,19 @@ import { openModal } from "@/src/reduxStore/states/modal";
 import { ModalEnum } from "@/src/types/shared/modal";
 import { selectOnAttributeEmbeddings } from "@/src/reduxStore/states/pages/settings";
 import { selectProjectId } from "@/src/reduxStore/states/project";
-import { createEvaluationSet, getSearchResults, getReformulationByQuestion, getPlaygroundQuestions, deleteQuestionFromHistory } from "@/src/services/base/playground";
+import { createEvaluationSet, getSearchResults, getReformulationByQuestion } from "@/src/services/base/playground";
 import { Embedding } from "@/src/types/components/projects/projectId/settings/embeddings";
 import KernDropdown from "@/submodules/react-components/components/KernDropdown";
-import { useState, useCallback, useEffect, useRef } from "react";
+import { useState, useCallback } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { RecordDisplay } from "@/src/components/shared/record-display/RecordDisplay";
 import { selectVisibleAttributesDataBrowser } from "@/src/reduxStore/states/pages/settings";
 import LoadingIcon from "@/submodules/react-components/components/LoadingIcon";
-import { IconFilterOff, IconFilter, IconCategoryPlus, IconLoader2, IconWand, IconHistory, IconTrash, } from '@tabler/icons-react'
+import { IconFilterOff, IconFilter, IconCategoryPlus, IconLoader2, IconWand } from '@tabler/icons-react'
 import PlaygroundSearchMetaFilterModal from "./PlaygroundSearchMetaFilterModal";
 import PlaygroundSearchReformulateModal from "./PlaygroundSearchReformulateModal";
 import { Tooltip } from "@nextui-org/react";
-import { PlaygroundQuestion } from "@/src/types/components/projects/projectId/settings/playground";
-import useOnClickOutside from "@/submodules/react-components/hooks/useHooks/useOnClickOutside";
+import QuestionHistory from "./QuestionHistory";
 
 const PLAYGROUND_LIMIT_DEFAULT = 10;
 const PLAYGROUND_THRESHOLD_DEFAULT = -9999;
@@ -34,15 +33,10 @@ export function PlaygroundSearch() {
     const [limit, setLimit] = useState(PLAYGROUND_LIMIT_DEFAULT);
     const [threshold, setThreshold] = useState(PLAYGROUND_THRESHOLD_DEFAULT);
     const [metaDataFilter, setMetaDataFilter] = useState(null);
-
     const [setCreationLoading, setSetCreationLoading] = useState(false);
     const [createdSet, setCreatedSet] = useState(false);
-    const [questionHistory, setQuestionHistory] = useState<PlaygroundQuestion[]>();
-    const [showHistory, setShowHistory] = useState(false);
     const [reformulationLoading, setReformulationLoading] = useState(false);
-
-    const dropdownRef = useRef(null);
-    useOnClickOutside(dropdownRef, () => { setShowHistory(false) });
+    const [refetchHistory, setRefetchHistory] = useState(false);
 
     const getSearchResultsPost = useCallback(() => {
         setLoading(true);
@@ -50,7 +44,10 @@ export function PlaygroundSearch() {
             setLoading(false);
             setCreatedSet(false);
             setSearchResults(result);
-            refetchQuestionMemory();
+            setRefetchHistory(true);
+            setTimeout(() => {
+                setRefetchHistory(false);
+            }, 1000);
         });
     }, [projectId, selectedEmbedding?.id, question, limit, metaDataFilter, threshold]);
 
@@ -70,11 +67,6 @@ export function PlaygroundSearch() {
         });
     }, [projectId, question, searchResults, createdSet]);
 
-    const handleHistoryClick = useCallback((entry: string) => {
-        setQuestion(entry);
-        setShowHistory(false);
-    }, []);
-
     const handleReformulation = useCallback((apiKey) => {
         setReformulationLoading(true);
         getReformulationByQuestion(projectId, question, apiKey, (result) => {
@@ -89,22 +81,6 @@ export function PlaygroundSearch() {
         )
     }, [projectId, question]);
 
-    useEffect(() => {
-        if (!projectId) return;
-        refetchQuestionMemory();
-    }, [projectId]);
-
-    function refetchQuestionMemory() {
-        getPlaygroundQuestions(projectId, (result) => {
-            setQuestionHistory(result);
-        });
-    }
-
-    function deleteQuestionFromHistoryFunc(questionId: string) {
-        deleteQuestionFromHistory(projectId, questionId, (result) => {
-            refetchQuestionMemory();
-        });
-    }
 
     return <div className="grid overflow-hidden grid-cols-2 h-[calc(100vh-200px)]">
         <div className="flex flex-col gap-y-2 m-3 h-full">
@@ -128,38 +104,7 @@ export function PlaygroundSearch() {
                             <IconWand className="h-5 w-5" />
                         </button>
                     </Tooltip>
-                    <Tooltip content="History" color="invert" placement="bottom">
-                        <button
-                            className="p-1 rounded-md text-gray-400 cursor-pointer hover:bg-gray-100 hover:text-gray-500 transition duration-150 ease-in-out"
-                            onClick={(e) => {
-                                e.preventDefault();
-                                e.stopPropagation();
-                                setShowHistory(!showHistory);
-                            }}
-                            onMouseDown={(e) => { e.preventDefault(); e.stopPropagation(); }}
-                        >
-                            <IconHistory className="h-5 w-5" />
-                        </button>
-                    </Tooltip>
-                    {showHistory && (
-                        <div ref={dropdownRef} className="absolute top-8 right-0 w-80 bg-gray-100 border border-gray-300 shadow-lg rounded-md p-2 z-50 max-h-96 overflow-y-auto">
-                            <div className="ml-1 text-md text-black mb-1">Last questions</div>
-                            <ul>
-                                {questionHistory && questionHistory.map((questionEntry, index) => (
-                                    <div key={questionEntry.id} className="p-2 flex items-center rounded-md border-2 border-white cursor-pointer transition duration-150 bg-white my-2 hover:border-black">
-                                        <li
-                                            className=" text-gray-700 "
-                                            onClick={() => handleHistoryClick(questionEntry.question)}
-                                        >
-                                            {questionEntry.question}
-                                        </li>
-                                        <IconTrash onClick={() => deleteQuestionFromHistoryFunc(questionEntry.id)}
-                                            className="h-6 w-6 text-red-700 cursor-pointer ml-auto" />
-                                    </div>
-                                ))}
-                            </ul>
-                        </div>
-                    )}
+                    <QuestionHistory setQuestion={(question: string) => setQuestion(question)} refetchHistory={refetchHistory} />
                 </div>
             </div>
             <div className="flex items-center justify-between">
